Authenticated added to pages
When a user wants to view a page that requires authentication, the frontend now asks the system whether a token exists and then asks Django whether this token is valid. Then the website asks every 60 seconds whether the token is still valid. <--- May change how we process this in the future.

- Will need to add token refresh schemes.
- Need to fix the problem with the views when fetching a specific club. Currently there is a bug in clubs.views.py file, I want the data to only load if the user is a valid user. But Django is always interpreting the user is AnonymousUser and hence preventing the use of the get and post methods.

What is particular interesting about this is that the Login and Logout views in accounts.views, use the exact same logic but Django is fine with this.