Events Added
There is now the functionality to create an event, start an event and mark the event as complete.

Currently events only have the ability to activate and deactivate members for that particular event. Next I will add the app for games in the Django backend.

This will give users 3 options:

- Peg generation of games (ie. people pick their game from the active members who aren't currently in a game)
- Skill based matchmaking generation, using elo I will devise an algorithm to: 1) make games consisting of people with the closest elo 2) make sure users get an even distribution of games
- Mathematically mixed games. Create an algorithm to ensure the games are as different as possible.