Game functionality added
In this update we improved the design of particular pages:
Logout
Login
Create Club
Events
Event view

We created the Elo app and introduced the logic for changing Elo when the game type is badminton.

For now auto-game create only works for same gender doubles, but can figure out fixes if there is not enough players for same gender doubles. The algorithm also effectively distributes even number of games.

In the next update we will improve on the algorithm, add an equivalent for mixed, so at the start of each game the algorithm will randomly pick between mixed or same gender then run the respective algorithm.

The Elo's are now accessible on the Account view page in React, but currently I have no implemented a way of displaying this data nicely. In the next few days I will make the account page more complicated:

Users will be able to see the sports they play
Users will be able to view the elo for those sports
Users will be able to visit a page specifically for them on that sport and view results from past games.