Major Update
The website has now gained features such as the ability to create DummyUsers, ie. users that are not accessible but can be used in games, this will be for people who want to play at the club but do not have the app or an account.

These dummyusers have both a dummyuser model and a customuser model. On delete both models get deleted. The reason for both models is so that later the customuser can be linked to a new customuser when that person registers.

EventView component has been cleaned up, soon I will add the ability to change modes. Currently there are two views that handle different game creation types: SBMM and Social mix.

I plan to make it easier to convert between these modes and add extra modes such as "Peg board" and make it possible to make all of these options SBMM.

The EventView now has a complete screen that displays the games, teams in the games and the corresponding score. The design of this needs to be improved.

A My clubs component has been added.

In the members list you can now make someone an Admin. I will be adding a de admin button later.

Likewise a uncomplete button for the event page will be added in case people accidently complete the game too early.